+++ ricaun.Nuke/Components/IGitRelease.cs
@@ -13,7 +13,7 @@ namespace ricaun.Nuke.Components
     /// <summary>
     /// IGitRelease
     /// </summary>
-    public interface IGitRelease : IRelease, IHazGitRepository, IHazGitVersion, IHazChangelog, INukeBuild
+    public interface IGitRelease : IRelease, IHazGitRepository, IHazGitVersion, IHazChangelog, IHazAssetRelease, INukeBuild
     {
         /// <summary>
         /// Target GitRelease
@@ -37,15 +37,15 @@ namespace ricaun.Nuke.Components
                     throw new Exception(errorMessage);
                 }
 
-                ReleaseGithubProject(project);
+                ReleaseGitHubProject(project);
             });
 
         /// <summary>
-        /// Release Github project with release notes
+        /// Release GitHub project with release notes
         /// </summary>
         /// <param name="project"></param>
         /// <param name="releaseAsPrerelease"></param>
-        void ReleaseGithubProject(Project project, bool releaseAsPrerelease = false)
+        void ReleaseGitHubProject(Project project, bool releaseAsPrerelease = false)
         {
             if (Directory.Exists(ReleaseDirectory) == false)
             {
@@ -74,10 +74,22 @@ namespace ricaun.Nuke.Components
                 return;
             }
 
+            var releaseNotes = GetReleaseNotes();
+            var releaseAssets = new ReleaseAssets
+            {
+                Project = project,
+                Version = version,
+                Notes = releaseNotes,
+                Files = releaseFiles,
+                Prerelease = releaseAsPrerelease
+            };
+
+            ReleaseAsset(releaseAssets);
+
             var newRelease = new Octokit.NewRelease(version)
             {
                 Name = version,
-                Body = GetReleaseNotes(),
+                Body = releaseNotes,
                 Draft = true,
                 TargetCommitish = GitVersion.Sha
             };