+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.9.0] / 2024-12-06
 ### Features
 - Enable sign files using `Azure Key Vault`.
+- Enable `IAssetRelease` to release assets before `GitRelease` and `GitPreRelease`.
 ### Build
 - Add `IAzureSignTool` to check if `AzureSignToolUtils` is installed.
 ### Updates